+++ src/server/routes/services/{serviceId}/resources/page.njk
@@ -22,36 +22,6 @@
       <div class="govuk-grid-column-full">
         {% call appSplitPane() %}
 
-          {% if pendingResourceRequestsRows %}
-            <section class="govuk-!-margin-bottom-9">
-              <span class="govuk-caption-l">Pending</span>
-
-              <h2 class="govuk-heading-l govuk-!-margin-bottom-2">Resource requests</h2>
-              <p class="govuk-!-margin-bottom-2">Resources pending creation for <strong>{{ entity.name }}</strong> </p>
-
-              {% call appPanel({
-                classes: "govuk-!-margin-bottom-6"
-              }) %}
-
-                {{ govukTable({
-                  classes: "app-table app-table--inverse govuk-!-margin-bottom-1",
-                  head: [
-                    { text: "Request generation workflow" },
-                    { text: "Request" },
-                    { text: "Requester" },
-                    { text: "Date" },
-                    { text: "Actions" }
-                  ],
-                  rows: pendingResourceRequestsRows
-                }) }}
-
-            {% endcall %}
-
-            </section>
-
-          {% endif %}
-
-
           <section>
             {{ appResourcesIntro({
               caption: "All",
