+++ src/server/routes/services/{serviceId}/resources/route.js
@@ -45,41 +45,13 @@ export default async function (request) {
 
   const environments = getEnvironments(request.auth.credentials?.scope)
 
-  const [resourcesPerEnv, pendingResourceRequests] = await Promise.all([
-    fetchResources(entity.name),
-    Promise.resolve([
-      {
-        requestedAt: '2026-07-09T08:43:26.793Z',
-        requestedBy: {
-          id: '',
-          displayName: 'David Beale'
-        },
-        workflow: {
-          workflow_run_id: 123,
-          run_url: '',
-          html_url: 'https://github.com/workflow'
-        },
-        pullRequest: { url: 'https://github.com/pr' },
-        resources: {
-          s3_buckets: [],
-          sqs_queues: [],
-          sns_topics: [],
-          sqs_sns_subscriptions: []
-        }
-      }
-    ])
-  ])
-
+  const resourcesPerEnv = await fetchResources(entity.name)
   if (!resourcesPerEnv) throw new Error('Failed to load resources')
 
   const rowsPerResourceType = transformResourcesToRows(
     environments,
     resourcesPerEnv
   )
-
-  const pendingResourceRequestsRows = pendingResourceRequests
-    ? transformPendingRequestsToRows(pendingResourceRequests)
-    : undefined
 
   const hasBuckets = rowsPerResourceType.s3_buckets?.length
 
@@ -96,7 +68,6 @@ export default async function (request) {
     entity,
     environments,
     tablesPerResourceType,
-    pendingResourceRequestsRows,
     hasBuckets,
     breadcrumbs: [
       {
@@ -170,14 +141,4 @@ function transformResourceRowsToTablesPerType(
         }
       ])
   )
-}
-
-function transformPendingRequestsToRows(pendingResourceRequests) {
-  return pendingResourceRequests.map((request) => [
-    { html: request.workflow.html_url },
-    { html: request.pullRequest.url },
-    { html: request.requestedBy.displayName },
-    { html: request.requestedAt },
-    { html: '' }
-  ])
 }
